test(grok): keep ICI-1393 live-stream tests under the 300-line cap
The regression file was 333 lines and failed ratchet. Move the fake
process and fixtures into support modules so the suite stays under
the limit without dropping mid-turn coverage.
